+++ platform/dashboard/src/pages/signup.tsx
@@ -78,7 +78,7 @@ export function SignUpPage() {
 
   if (platformSettings.loading) {
     return (
-      <AuthLayout title={'Sign in to RushDB'}>
+      <AuthLayout title={'Sign in to RushDB'} type="signin">
         <div className="m-auto flex content-center items-center justify-between">
           <Spinner />
         </div>
@@ -87,7 +87,7 @@ export function SignUpPage() {
   }
 
   return (
-    <AuthLayout title={'Create new RushDB account'}>
+    <AuthLayout title={'Create new RushDB account'} type="signup">
       {hasOauth ?
         <>
           <div className="flex w-full justify-between gap-2">