+++ platform/dashboard/CHANGELOG.md
@@ -1,5 +1,58 @@
 # rushdb-dashboard
 
+## 0.17.1
+
+### Patch Changes
+
+- f0a395c: SEO Optimizations
+- Updated dependencies [f0a395c]
+  - @rushdb/javascript-sdk@0.17.1
+
+## 0.17.0
+
+### Minor Changes
+
+- 1cdfbf7: Implemented RawApi mode for the dashboard and minor fixes
+
+### Patch Changes
+
+- 29940a0: Fix lock file
+- Updated dependencies [29940a0]
+- Updated dependencies [1cdfbf7]
+  - @rushdb/javascript-sdk@0.17.0
+
+## 0.16.0
+
+### Minor Changes
+
+- 57a566b: Update LP & Relationships UI and Query
+
+### Patch Changes
+
+- Updated dependencies [57a566b]
+  - @rushdb/javascript-sdk@0.16.0
+
+## 0.15.0
+
+### Minor Changes
+
+- 6aa0dca: Update docs
+- bcdf10d: Add llms.txt and llms-full.txt to docs generation
+
+### Patch Changes
+
+- Updated dependencies [6aa0dca]
+- Updated dependencies [bcdf10d]
+  - @rushdb/javascript-sdk@0.15.0
+
+## 0.14.1
+
+### Patch Changes
+
+- 7bfd0aa: Make label field required in import modal in dashboard
+- Updated dependencies [7bfd0aa]
+  - @rushdb/javascript-sdk@0.14.1
+
 ## 0.14.0
 
 ### Minor Changes